entrance: widen EntranceInfo.spawn to s16
The field beside the one 973f56be widened, and the same trap: an s8 holding an
index into a table that is only going to grow. Less urgent than `scene` was -
every custom scene this project makes uses spawn 0, and a scene needs 128 spawn
points before it bites - but it costs nothing, because widening `scene` had
already opened the padding it needs. The struct stays 6 bytes.

Widening it removes the truncation, not the ceiling. Play_SpawnScene hands the
value to PlayState.curSpawn, a u8, which is what actually indexes
setupEntranceList - so a spawn index now stops at 255 rather than 127. curSpawn
is left alone: raising it further needs a scene with more than 255 spawn points,
and vanilla's largest has 13.

Entrance_SceneAndSpawnAre follows. Its whole job is comparing against
EntranceInfo, so u8 parameters were the same silent truncation one level over -
a caller passing a scene id past 255 would have matched the wrong entrance
rather than none. Every call site passes a vanilla SCENE_* constant, so nothing
changes today.

sturdy-bassoon#28

+++ soh/soh/Enhancements/randomizer/randomizer_entrance.c
@@ -301,7 +301,11 @@ s16 Entrance_OverrideDynamicExit(s16 dynamicExitIndex) {
     return Grotto_OverrideSpecialEntrance(Entrance_GetOverride(dynamicExitList[dynamicExitIndex]));
 }
 
-u32 Entrance_SceneAndSpawnAre(u8 scene, u8 spawn) {
+// `scene` and `spawn` are s16 to match the EntranceInfo fields they are compared against, widened
+// in sturdy-bassoon#27 and #28. As u8s this was the same silent truncation one level over: a
+// caller passing a scene id past 255 would have matched the wrong entrance rather than none.
+// Every call site passes a vanilla SCENE_* constant, so nothing here changes today.
+u32 Entrance_SceneAndSpawnAre(s16 scene, s16 spawn) {
     s16 entranceIndex;
 
     // Adjust the entrance to account for the exact scene/spawn combination for child/adult and day/night
