+++ soh/soh/Enhancements/randomizer/randomizer_entrance.h
@@ -40,7 +40,7 @@ int16_t Entrance_GetOverride(int16_t index);
 int16_t Entrance_OverrideNextIndex(int16_t nextEntranceIndex);
 int16_t Entrance_PeekNextIndexOverride(int16_t nextEntranceIndex);
 int16_t Entrance_OverrideDynamicExit(int16_t dynamicExitIndex);
-uint32_t Entrance_SceneAndSpawnAre(uint8_t scene, uint8_t spawn);
+uint32_t Entrance_SceneAndSpawnAre(int16_t scene, int16_t spawn);
 void Entrance_SetGameOverEntrance(void);
 void Entrance_SetSavewarpEntrance(void);
 void Entrance_SetWarpSongEntrance(void);
